+++ src/main/java/com/room414/racingbets/dal/abstraction/dao/HorseDao.java
@@ -11,11 +11,11 @@ import java.util.List;
  * @author Alexander Melashchenko
  * @version 1.0 26 Feb 2017
  */
-public interface HorseDao extends CrudDao<Integer, Horse> {
+public interface HorseDao extends CrudDao<Long, Horse> {
     /**
      * @return List of Horses whose name starts with namePart or empty list if no found.
      */
-    List<Horse> findByNamePart(String namePart, int offset, int limit);
+    List<Horse> findByNamePart(String namePart, long offset, long limit);
 
     /**
      * @return count of Horses whose name starts with namePart
